fix: change weight from int to float

+++ db/migrate/20240729163918_create_models.rb
@@ -3,7 +3,7 @@ class CreateModels < ActiveRecord::Migration[7.1]
     create_table :models do |t|
       t.string :name, null: false
       t.string :overview, null: false
-      t.integer :weight, null: false
+      t.float :weight, null: false
       t.integer :heel_to_toe_drop, null: false
       t.integer :iteration, null: false
       t.boolean :apma_accepted, null: false
